Major Performance Improvements
 * Instead of generating a new string/parameter vector by
inserting/deleting each possible mutation into a new template, simply
use offset indexing to maintain the current template and return the
"new" values when requested.  Several new methods to support this,
called "ApplyVirtualMutation, etc".  Might refactor this later to make
it more transparent.

 * Have all subreads share the same underyling template/parameter pair
to avoid having multiple copies of the same data.  I introduced the
WrappedTemplateParameterPair class as a thin layer over the underyling
data taht adjusts for individual reads alignment spans.

 * Remove abstract classes that only had one subclass.  Makes using
templates harder in C#, but removes unneeded complexity.

* Removed several classes that aren’t used but caused compiler problems.

Needs more work to:
	* Generate far more tests
	* Remove branching conditions

+++ src/ConsensusCore/src/C++/Matrix/DenseMatrix-inl.hpp
@@ -72,13 +72,13 @@ namespace ConsensusCore {
     inline const int
     DenseMatrix::Rows() const
     {
-        return size1();
+        return (int)size1();
     }
 
     inline const  int
     DenseMatrix::Columns() const
     {
-        return size2();
+        return (int)size2();
     }
 
     //
